// test/2/src/type.ts
export type Iconfig = IconfigWs | IconfigWsReverse;

export interface IconfigWs {
	mode: 'ws';
	address: string;
	port: number;
	retry: number;
}

export interface IconfigWsReverse {
	mode: 'ws-reverse';
	port: number;
}

export interface BotInfo {
	self_id: number;
	connect: number;
	heartbeat: number;
	status: EventStatusType;
}

/* Api */
export type Msg = string | Message | Message[];
export type MessageCqType =
	| 'face'
	| 'record'
	| 'video'
	| 'at'
	| 'share'
	| 'music'
	| 'image'
	| 'reply'
	| 'redbag'
	| 'poke'
	| 'gift'
	| 'forward'
	| 'node'
	| 'xml'
	| 'json'
	| 'cardimage'
	| 'tts';
export type MessageDataType =
	| MessageFace
	| MessageRecord
	| MessageVideo
	| MessageAt
	| MessageShare
	| MessageMusic
	| MessageImage
	| MessageReply
	| MessageRedbag
	| MessagePoke
	| MessageGift
	| MessageForward
	| MessageNode
	| MessageXml
	| MessageJson
	| MessageCardImage
	| MessageTts;
export interface Message {
	type: MessageCqType;
	data: MessageDataType;
}
// ...
export type EventPostType = 'message' | 'message_sent' | 'request' | 'notice' | 'meta_event';
export type EventMessageType = 'private' | 'group';
export type EventSubType =
	| 'private'
	| 'normal'
	| 'anonymous'
	| 'group_self'
	| 'group'
	| 'notice'
	| 'connect'
	| 'approve'
	| 'add'
	| 'invite'
	| 'leave'
	| 'kick'
	| 'kick_me'
	| 'set'
	| 'unset'
	| 'ban'
	| 'lift_ban';
export type EventRequestType = 'private' | 'group';
export type EventNoticeType =
	| 'group_upload'
	| 'group_admin'
	| 'group_decrease'
	| 'group_increase'
	| 'group_ban'
	| 'private_add'
	| 'group_recall'
	| 'private_recall'
	| 'group_card'
	| 'offline_file'
	| 'client_status'
	| 'essence'
	| 'notify';
export type EventNoticeNotifySubType = 'honor' | 'poke' | 'lucky_king' | 'title';
export type EventMetaEventType = 'lifecycle' | 'heartbeat';

export type EvenetSenderRoleType = 'owner' | 'admin' | 'member';

export interface EventSenderType {
	user_id: number;
	nickname: string;
	sex: eventDataMsgSenderSex;
	age: number;
	group_id?: number;
	card?: string;
	area?: string;
	level?: string;
	role?: EvenetSenderRoleType;
	title?: string;
}

export interface EventStatusType {
	app_initialized: boolean;
	app_enabled: boolean;
	plugins_good: boolean | null;
	app_good: boolean;
	online: boolean;
	stat: EventStatType;
}

export interface EventStatType {
	packet_received: number;
	packet_sent: number;
	packet_lost: number;
	message_received: number;
	message_sent: number;
	disconnect_times: number;
	lost_times: number;
	last_message_time: number;
}

export interface EventDataType extends obj {
	post_type: EventPostType;
	message_type: EventMessageType;
	sub_type: EventSubType;
	request_type: EventRequestType;
	notice_type: EventNoticeType;
	notice_notify_subtype: EventNoticeNotifySubType;
	meta_event_type: EventMetaEventType;
	time: number;
	self_id: number;
	message_: Message;
	message: string;
	message_id: number;
	raw_message: string;
	user_id: number;
	group_id?: number;
	operator_id?: number;
	duration?: number;
	sender: EventSenderType;
	status: EventStatusType;
	font: number | 0;
}